pgp_server/requirements.txt
python-gnupg==0.5.2
pika==1.3.2
python-dotenv==1.0.1
